Extract default 'publish_schema' action into explicit task (#396)
'publish_schema' used to be a default action on every prepare definition
that was executed in the 'complete_prepare_process' method, called when
all tasks were complete.

However, we also need to execute this step from Airflow. This default
step is now extracted into its own action. A '*' dependency is added to
denote steps that need to happen after all other steps. The
_create_tasks method will take this into account and set the necessary
dependencies.