Add citus columnar support to Postgres actions (#399)
Added two util functions for generating CREATE TABLE statements for
Postgres.

These functions will check if the citus extension is available on the
database and whether we're on the correct Postgres version. If so, it
will return the statements to create columnar tables. Otherwise it will
return the default Postgres CREATE TABLE statement.

Note that citus can be enabled on Postgres < 12 as well, but this is a
different and more cumbersome implementation to work with. This would
need to be implemented in a very different way, hence the check for PG
version >= 12.

PR also contains LOADS of updates on the existing two BRK and BRK2
prepare pipelines; Citus does not support many UPDATE actions, so the
pipelines needed to be rearranged. Did not see this coming, hence this
unexpectedly large PR. Apologies for that!